moves rest of exported backgroudn contextMenu functions to separate files
